Move the DefaultPrimaryButton into the PrimaryButton
