+++ stripe/res/layout/activity_payment_sheet.xml
@@ -63,7 +63,7 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent">
-            <com.stripe.android.paymentsheet.DefaultPrimaryButton
+            <com.stripe.android.paymentsheet.ui.PrimaryButton
                 android:id="@+id/buy_button"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
