+++ stripe/res/layout/primary_button.xml
