+++ stripe/res/layout/stripe_activity_payment_options.xml
@@ -52,7 +52,7 @@
                 style="@style/StripePaymentSheetErrorMessage"/>
         </LinearLayout>
 
-        <com.stripe.android.paymentsheet.DefaultPrimaryButton
+        <com.stripe.android.paymentsheet.ui.PrimaryButton
             android:id="@+id/add_button"
             android:layout_marginTop="@dimen/stripe_paymentsheet_button_container_spacing"
             android:layout_marginStart="@dimen/stripe_paymentsheet_outer_spacing_horizontal"
